requirements.txt
gradio
torch
transformers
accelerate
safetensors
huggingface_hub
tensorflow-cpu
Pillow
numpy